+++ influxql/scanner.go
@@ -53,6 +53,12 @@ func (s *Scanner) Scan() (tok Token, pos Pos, lit string) {
 			return s.scanNumber()
 		}
 		return DOT, pos, ""
+	case '$':
+		tok, _, lit := s.scanIdent()
+		if tok == IDENT {
+			tok = BOUNDPARAM
+		}
+		return tok, pos, lit
 	case '+', '-':
 		return s.scanNumber()
 	case '*':